ALEA-940: Changed the San Antonio VS code task to always run on x86 on osx https://gitlab.com/alea-2020/command-data-handling/obc2-firmware/-/merge_requests/181
# Overview
Changed the vscode tasks to prepend the San Antonio run command with `arch -x86_64`.

Additionally, the docs have been changed in the ALEASAT wiki to change the install instructions to include the same considerations for virtualization.

# Testing
No testing is needed since this shouldn't affect any other components of the firmware.