ALEA-977 Add uptime command https://gitlab.com/alea-2020/command-data-handling/obc2-firmware/-/merge_requests/206
# Overview
Creates GET_UPTIME and GET_RTC_EPOCH commands using the RTI free running counter 0 and the RTC respectively. Additionally, adds a quick command to san-antonio to access the uptime 

It is worth noting the the GET_RTC_EPOCH are in backup_epoch.c since it relies on the internal of backup_epoch. There are only 2 free-running counters and since the backup epoch is running at an appropriate frequency this shouldn't be an issue. When backup_epoch is set it doesn't change the underlying counter, rather it changes an offset value.

# Testing
Tested locally on Launchpad, there is an issue with the discrepancy between the RTC and RTC using GET_UPTIME and GET_RTC_EPOCH. I will let them both run but it looks like after about an hour there is a ~5 second discrepancy between the RTC and the RTI.

# Reviewers
@EthanJoyce @noah.tajwar.orbit @Julian-Mentasti


# Checklist

## Before Review (Submitter)
* [X] This branch has been rebased on top of the latest version of `dev`.
* [X] I have done a self-review of this MR, and left comments on anything that may be unclear
* [X] I have added unit & hill tests to any code that can be tested

## After Review, Before Merge (Submitter)
* [ ] Relevant design docs and wiki pages have been updated.
* [ ] Issue has been closed, if appropriate.
* [ ] Changes from review have been tested on OBC and LaunchPad