ALEA-2364 Fixed Chained PreCompute Performance https://gitlab.com/alea-2020/command-data-handling/obc2-firmware/-/merge_requests/274
Parent issue: PP#2364

# Overview
OrbitDynamicsModel pre-computes data in batches. Previously EarthMagneticFieldModel was un-batching the orbit dynamics data and doing it's pre-compute on one element at a time which was created a lot of IPC overhead. Now there's an option to not un-batch the pre-compute data so you can reduce IPC.

## Profiling

Test Case: Propagate orbit and calculate magnetic field up to 54000 seconds.

### Original Code

![image](/uploads/9efa759885495018ec8264a75f154402/image.png){width=1033 height=536}

### Without Un-Batching

![image](/uploads/77a9bf1488b6cf68fe5059050bef5069/image.png){width=1029 height=528}
![image](/uploads/3439ea9f77fe51e98c9913cb3ced4810/image.png){width=1026 height=522}

# Testing
- Unit Tests Passed
- Profiled (see above)

# Reviewers
@yel-wishahy